refactor(code-actions): split provider routing modules (#8197)
